requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp_socks==0.10.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.2
Brotli==1.1.0
browser-cookie3==0.20.1
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.1
cssselect2==0.8.0
curl_cffi==0.9.0
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.4.0
Django==5.2.4
django-formtools==2.5.1
duckduckgo_search==7.5.1
EbookLib==0.18
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
g4f==0.4.8.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.1
lz4==4.4.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mss==10.0.0
multidict==6.1.0
narwhals==1.29.1
nest-asyncio==1.6.0
nodriver==0.40
numpy==2.2.3
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plyer==2.1.0
primp==0.14.0
propcache==0.3.0
protobuf==5.29.3
proxy_tools==0.1.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
python-socks==2.7.1
pythonnet==3.0.5
pytz==2025.1
pywebview==5.4
pywin32==308
qrcode==7.4.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
shadowcopy==0.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.43.1
tenacity==9.0.0
termtype==1.0.4
termtyper==2.0.5
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.32.0
watchdog==6.0.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
windows-curses==2.3.1
WMI==1.5.1
wrapt==1.17.2
yarl==1.18.3
